refactor: move services to shared-ui
