+++ packages/mask/dashboard/components/Restore/RestoreFromCloud/ConfirmBackupInfo.tsx
@@ -2,7 +2,7 @@ import { decryptBackup } from '@masknet/backup-format'
 import { decode, encode } from '@msgpack/msgpack'
 import { Box } from '@mui/material'
 import { memo, useCallback, useLayoutEffect, useState } from 'react'
-import { Services } from '../../../API.js'
+import { Services } from '../../../../shared-ui/service.js'
 import { usePersonaRecovery } from '../../../contexts/index.js'
 import { useDashboardI18N } from '../../../locales/index.js'
 import { fetchBackupValue } from '../../../pages/Settings/api.js'
