+++ packages/mask/dashboard/components/Restore/RestoreFromCloud/index.tsx
@@ -5,7 +5,7 @@ import { Box } from '@mui/material'
 import { DashboardRoutes } from '@masknet/shared-base'
 import { useCustomSnackbar } from '@masknet/theme'
 import { WalletServiceRef } from '@masknet/plugin-infra/dom'
-import { Services } from '../../../API.js'
+import { Services } from '../../../../shared-ui/service.js'
 import { useDashboardI18N } from '../../../locales/index.js'
 import { PersonaContext } from '../../../pages/Personas/hooks/usePersonaContext.js'
 import { BackupPreview } from '../../../pages/Settings/components/BackupPreview.js'
