+++ packages/mask/dashboard/components/Restore/RestorePersonaFromLocal.tsx
@@ -8,7 +8,7 @@ import { FileFrame, UploadDropArea } from '@masknet/shared'
 import { makeStyles, useCustomSnackbar } from '@masknet/theme'
 import { decode, encode } from '@msgpack/msgpack'
 import { WalletServiceRef } from '@masknet/plugin-infra/dom'
-import { Services } from '../../API.js'
+import { Services } from '../../../shared-ui/service.js'
 import { usePersonaRecovery } from '../../contexts/RecoveryContext.js'
 import { useDashboardI18N } from '../../locales/index.js'
 import { BackupPreview } from '../../pages/Settings/components/BackupPreview.js'
