+++ packages/mask/dashboard/entry.tsx
@@ -1,3 +1,3 @@
 import './initialization/i18n.js'
 export { IntegratedDashboard } from './Dashboard.js'
-export { setService, setMessages } from './API.js'
+export { setMessages } from './API.js'
