+++ packages/mask/dashboard/pages/CreateMaskWallet/Onboarding/index.tsx
@@ -8,7 +8,7 @@ import { Icons } from '@masknet/icons'
 import { Trend } from '../../../assets/index.js'
 import { PopupRoutes } from '@masknet/shared-base'
 
-import { Services } from '../../../API.js'
+import { Services } from '../../../../shared-ui/service.js'
 import { OnboardingWriter } from '../../../components/OnboardingWriter/index.js'
 
 const useStyles = makeStyles()((theme) => ({
