+++ packages/mask/dashboard/pages/CreateMaskWallet/context.ts
@@ -5,7 +5,7 @@ import { getDefaultWalletPassword, CrossIsolationMessages, PopupRoutes } from '@
 import { Web3 } from '@masknet/web3-providers'
 import { ProviderType } from '@masknet/web3-shared-evm'
 import { WalletServiceRef } from '@masknet/plugin-infra/dom'
-import { Services } from '../../API.js'
+import { Services } from '../../../shared-ui/service.js'
 
 function useContext() {
     const wallets = useWallets()
