+++ packages/mask/dashboard/pages/Personas/components/ContactTableRow/index.tsx
@@ -6,7 +6,7 @@ import { Icons } from '@masknet/icons'
 import { LoadingButton } from '@mui/lab'
 import type { RelationProfile } from '@masknet/shared-base'
 import { Image, Icon } from '@masknet/shared'
-import { Services } from '../../../../API.js'
+import { Services } from '../../../../../shared-ui/service.js'
 import { useDashboardI18N } from '../../../../locales/index.js'
 import { useAddContactToFavorite, useRemoveContactFromFavorite } from '../../hooks/useFavoriteContact.js'
 import { PersonaContext } from '../../hooks/usePersonaContext.js'
