+++ packages/mask/dashboard/pages/Personas/components/DeletePersonaDialog/index.tsx
@@ -2,7 +2,7 @@ import { Button, DialogActions, DialogContent, Typography } from '@mui/material'
 import { memo, useCallback } from 'react'
 import { MaskColorVar, MaskDialog, useCustomSnackbar } from '@masknet/theme'
 import { DashboardTrans, useDashboardI18N } from '../../../../locales/index.js'
-import { Services } from '../../../../API.js'
+import { Services } from '../../../../../shared-ui/service.js'
 import { type PersonaIdentifier, DashboardRoutes } from '@masknet/shared-base'
 import { PersonaContext } from '../../hooks/usePersonaContext.js'
 import { useNavigate } from 'react-router-dom'
