+++ packages/mask/dashboard/pages/Personas/components/LogoutPersonaDialog/index.tsx
@@ -10,7 +10,7 @@ import { ManageWallet } from '@masknet/shared'
 import { useWallets } from '@masknet/web3-hooks-base'
 import { formatEthereumAddress, ProviderType } from '@masknet/web3-shared-evm'
 import { WalletServiceRef } from '@masknet/plugin-infra/dom'
-import { Services } from '../../../../API.js'
+import { Services } from '../../../../../shared-ui/service.js'
 import { DashboardTrans, useDashboardI18N } from '../../../../locales/index.js'
 import { PersonaContext } from '../../hooks/usePersonaContext.js'
 import PasswordField from '../../../../components/PasswordField/index.js'
