+++ packages/mask/dashboard/pages/Personas/hooks/useExportMnemonicWords.ts
@@ -1,6 +1,6 @@
 import { useAsync } from 'react-use'
 import type { PersonaIdentifier } from '@masknet/shared-base'
-import { Services } from '../../../API.js'
+import { Services } from '../../../../shared-ui/service.js'
 import type { AsyncState } from 'react-use/lib/useAsyncFn.js'
 
 export function useExportMnemonicWords(identifier: PersonaIdentifier): AsyncState<string> {
