+++ packages/mask/dashboard/pages/Settings/components/AppearanceSetting.tsx
@@ -1,6 +1,6 @@
 import { MenuItem, type SelectChangeEvent } from '@mui/material'
 import { Appearance } from '@masknet/public-api'
-import { Services } from '../../../API.js'
+import { Services } from '../../../../shared-ui/service.js'
 import { useAppearance } from '../../Personas/api.js'
 import SettingSelect from './SettingSelect.js'
 import { useDashboardI18N } from '../../../locales/index.js'
