+++ packages/mask/dashboard/pages/Settings/components/dialogs/RestoreDialog.tsx
@@ -4,7 +4,7 @@ import { TabContext, TabList, TabPanel, tabPanelClasses } from '@mui/lab'
 import { InputBase, Tab, inputBaseClasses, styled, tabClasses, tabsClasses } from '@mui/material'
 import { useState } from 'react'
 import { useAsync } from 'react-use'
-import { Services } from '../../../../API.js'
+import { Services } from '../../../../../shared-ui/service.js'
 import ConfirmDialog from '../../../../components/ConfirmDialog/index.js'
 import FileUpload from '../../../../components/FileUpload/index.js'
 import { BackupPreview } from '../BackupPreview.js'
