+++ packages/mask/dashboard/pages/SetupPersona/Mnemonic/index.tsx
@@ -7,7 +7,7 @@ import { toBlob } from 'html-to-image'
 import { memo, useCallback, useRef, useState } from 'react'
 import { useLocation, useNavigate } from 'react-router-dom'
 import { useAsync, useAsyncFn } from 'react-use'
-import { Services } from '../../../API.js'
+import { Services } from '../../../../shared-ui/service.js'
 import { PrimaryButton } from '../../../components/PrimaryButton/index.js'
 import { SetupFrameController } from '../../../components/SetupFrame/index.js'
 import { useCreatePersonaV2 } from '../../../hooks/useCreatePersonaV2.js'
