+++ packages/mask/dashboard/pages/SetupPersona/Welcome/index.tsx
@@ -4,7 +4,7 @@ import { Checkbox, FormControlLabel, Typography } from '@mui/material'
 import { memo, useCallback, useState } from 'react'
 import { useNavigate, useSearchParams } from 'react-router-dom'
 import urlcat from 'urlcat'
-import { Services } from '../../../API.js'
+import { Services } from '../../../../shared-ui/service.js'
 import { TermsAgreedContext } from '../../../hooks/useTermsAgreed.js'
 import { DashboardTrans, useDashboardI18N } from '../../../locales/index.js'
 import { SecondaryButton } from '../../../components/SecondaryButton/index.js'
