+++ packages/mask/dashboard/utils/kv-storage.ts
@@ -1,5 +1,6 @@
 import { createKVStorageHost, type KVStorageBackend } from '@masknet/shared-base'
-import { Services, Messages } from '../API.js'
+import { Messages } from '../API.js'
+import { Services } from '../../shared-ui/service.js'
 
 const memory: KVStorageBackend = {
     beforeAutoSync: Promise.resolve(),
